# OpenApi/cloudboard-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CloudBoard.ApiService | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sorting-applications": {
      "get": {
        "tags": [
          "CloudBoard.ApiService"
        ],
        "operationId": "GetAllSortingApplications",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SortingApplicationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CloudBoard.ApiService"
        ],
        "operationId": "CreateSortingApplication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SortingApplicationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SortingApplicationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/sorting-applications/{applicationId}": {
      "get": {
        "tags": [
          "CloudBoard.ApiService"
        ],
        "operationId": "GetSortingApplicationById",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SortingApplicationDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "CloudBoard.ApiService"
        ],
        "operationId": "UpdateSortingApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SortingApplicationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SortingApplicationDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "CloudBoard.ApiService"
        ],
        "operationId": "DeleteSortingApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/sorting-applications/{applicationId}/process-steps": {
      "get": {
        "tags": [
          "CloudBoard.ApiService"
        ],
        "operationId": "GetProcessStepsBySortingApplicationId",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessStepDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MarketSegmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "segmentCode": {
            "type": "string"
          },
          "segmentName": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "businessUnit": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ProcessStepDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "processStepId": {
            "type": "string"
          },
          "stepCode": {
            "type": "string"
          },
          "processStepName": {
            "type": "string"
          },
          "stepType": {
            "type": "string"
          },
          "infeedMaterialDescription": {
            "type": "string"
          },
          "mainMaterialForEjection": {
            "type": "string"
          },
          "conflictingMaterials": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "marketSegment": {
            "$ref": "#/components/schemas/MarketSegmentDto"
          },
          "targetMaterials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetMaterialDto"
            }
          }
        }
      },
      "SortingApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "processSteps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessStepDto"
            }
          }
        }
      },
      "TargetMaterialDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "materialCode": {
            "type": "string"
          },
          "materialName": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "form": {
            "type": "string"
          },
          "isContaminant": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}